Added support for dynamically loading AdminX apps
ref https://github.com/TryGhost/ActivityPub/pull/1161

The proposed versioning strategy for the ActivityPub client relies on us
being able to dynamically fetch the CDN url from the backend. We're
unable to do that in the asset-delivery module, as this is run during
compile time, so instead the AdminX Component has been updated to
conditionaly, and safely, attempt to fetch the remote config url, if
present, and use the cdn url from that.